Jewelry Crafting and Summerset Crafted Set ItemIDs
New material requirements table for Jewelry, which does not follow the pattern that blacksmithing/clothing/woodworking all do.
New skill/ability table for improvement skills, since jewelry (once again) does not follow the blacksmithing/clothing/woodworking pattern.

Gratuitous refactor of enoughMaterials() to change from deeply-nested-if-chain to a gauntlet-of-if-blocks so that I could keep the "what you're out of" message close to the test that detects it. Specifically, injecting the "don't look for style materials on jewelry" logic into this deeply-nested-if-chain broke Zig's brain.