Minor updates
Fixed a bug where choosing an enchanting level for glyphs or to craft would cause errors if it was not a valid enchanting level. Now, it'll just get the level just before (i.e. if you choose lvl 12 it'll give you lvl 10)
Note: Also went to 3.0 to stop having so many 9s
Fixed a bug where sometimes New Moon Acolyte gear would not be crafted
Fixed a bug where if LibStub was not running there would be an error
Fixed a bug where if a non auto craft item's base item was crafted it would not improve it automatically
Fixed backwards compatability bug
Fixed a bug where there would sometimes be a lua error in the get materials